refactor : 불필요 파일 삭제, 파일 경로 수정
