--- src/app/Providers.tsx
@@ -1,6 +1,0 @@
-"use client";
-import { SessionProvider } from "next-auth/react";
-
-export function Providers({ children }: { children: React.ReactNode }) {
-  return <SessionProvider>{children}</SessionProvider>;
-}
