+++ src/app/components/PlaylistBar.tsx
